Implemented tokenised method to delete and add product with an globally made function . in admin panel delete and add product section has been tokenised

+++ backend/productRouter.js
@@ -3,12 +3,10 @@ import productModel from "./DB/productModel.js";
 import multer from "multer";
 import path from "path"
 import Jwt from "jsonwebtoken";
+import Authorization from "./Authorisation.js";
 
 const productRouter = express.Router();
 const secretKey = 'ajay-shekhawat'
-// const verifyToken =()=>{
-
-// }
 
 productRouter.get("/", async (req, res) => {
     let allProducts = await productModel.find({});
@@ -28,33 +26,33 @@ const upload = multer({ storage: storage });
 
 productRouter.post("/add", upload.single("image"), async (req, res) => {
     const token = req.headers.authorization
-    console.log(token);
-    if (!token) {
-        return res.status(401).json({ message: 'Unauthorized' });
-    }
-    Jwt.verify(token, secretKey, async (err, decoded) => {
-        if (err) {
-            console.log("errr wala chla code exuecute hua",err);
-            return res.json({ message: 'Token is invalid' });
-            // return res.status(401).json({ message: 'Token is invalid' });
-        }
-        console.log(decoded);
-        console.log("Token match hogya");
+    const AuthorisationResult = await Authorization(token, secretKey)
+    if (AuthorisationResult) {
         const { name, price, category, company } = req.body;
         const image = req.file;
         let productToAdd = new productModel({ name, price, category, company, image })
         let result = await productToAdd.save();
         res.json(result)
-    })
+    } else {
+        return res.json({ message: 'Token is invalid' });
+    }
 })
 
 
 // To Delete
 productRouter.delete("/del/:id", async (req, res) => {
+    const token = req.headers.authorization
     const productToDeleteId = req.params.id;
-    let result = await productModel.deleteOne({ _id: productToDeleteId })
-    res.json(result);
+    const AuthorisationResult = await Authorization(token, secretKey)
+    if (AuthorisationResult) {
+        let result = await productModel.deleteOne({ _id: productToDeleteId })
+        res.json(result);
+    } else {
+        res.json({ message: 'Token is invalid' })
+    }
 })
+
+
 //To Update
 productRouter.get("/edit/:id", async (req, res) => {
     let result = await productModel.findOne({ _id: req.params.id })
